Break each endpoint in index.js into its own file.
